+++ src/extensions/surveys/components/QuestionTypes.tsx
@@ -24,10 +24,12 @@ import { Cancel, QuestionHeader } from './QuestionHeader'
 
 export function OpenTextQuestion({
     question,
+    forceDisableHtml,
     appearance,
     onSubmit,
 }: {
     question: BasicSurveyQuestion
+    forceDisableHtml: boolean
     appearance: SurveyAppearance
     onSubmit: (text: string) => void
 }) {
@@ -45,7 +47,9 @@ export function OpenTextQuestion({
             <QuestionHeader
                 question={question.question}
                 description={question.description}
+                descriptionContentType={question.descriptionContentType}
                 backgroundColor={appearance.backgroundColor}
+                forceDisableHtml={forceDisableHtml}
             />
             <textarea rows={4} placeholder={appearance?.placeholder} onInput={(e) => setText(e.currentTarget.value)} />
             <BottomSection
@@ -60,10 +64,12 @@ export function OpenTextQuestion({
 
 export function LinkQuestion({
     question,
+    forceDisableHtml,
     appearance,
     onSubmit,
 }: {
     question: LinkSurveyQuestion
+    forceDisableHtml: boolean
     appearance: SurveyAppearance
     onSubmit: (clicked: string) => void
 }) {
@@ -72,7 +78,12 @@ export function LinkQuestion({
     return (
         <div className="survey-box">
             <Cancel onClick={() => handleCloseSurveyPopup()} />
-            <QuestionHeader question={question.question} description={question.description} />
+            <QuestionHeader
+                question={question.question}
+                description={question.description}
+                descriptionContentType={question.descriptionContentType}
+                forceDisableHtml={forceDisableHtml}
+            />
             <BottomSection
                 text={question.buttonText || 'Submit'}
                 submitDisabled={false}
@@ -86,11 +97,13 @@ export function LinkQuestion({
 
 export function RatingQuestion({
     question,
+    forceDisableHtml,
     displayQuestionIndex,
     appearance,
     onSubmit,
 }: {
     question: RatingSurveyQuestion
+    forceDisableHtml: boolean
     displayQuestionIndex: number
     appearance: SurveyAppearance
     onSubmit: (rating: number | null) => void
@@ -106,6 +119,8 @@ export function RatingQuestion({
             <QuestionHeader
                 question={question.question}
                 description={question.description}
+                descriptionContentType={question.descriptionContentType}
+                forceDisableHtml={forceDisableHtml}
                 backgroundColor={appearance.backgroundColor}
             />
             <div className="rating-section">
@@ -207,11 +222,13 @@ export function RatingButton({
 
 export function MultipleChoiceQuestion({
     question,
+    forceDisableHtml,
     displayQuestionIndex,
     appearance,
     onSubmit,
 }: {
     question: MultipleSurveyQuestion
+    forceDisableHtml: boolean
     displayQuestionIndex: number
     appearance: SurveyAppearance
     onSubmit: (choices: string | string[] | null) => void
@@ -236,6 +253,8 @@ export function MultipleChoiceQuestion({
             <QuestionHeader
                 question={question.question}
                 description={question.description}
+                descriptionContentType={question.descriptionContentType}
+                forceDisableHtml={forceDisableHtml}
                 backgroundColor={appearance.backgroundColor}
             />
             <div className="multiple-choice-options">