+++ src/posthog-surveys-types.ts
@@ -21,6 +21,7 @@ export interface SurveyAppearance {
     displayThankYouMessage?: boolean
     thankYouMessageHeader?: string
     thankYouMessageDescription?: string
+    thankYouMessageDescriptionContentType?: SurveyQuestionDescriptionContentType
     borderColor?: string
     position?: 'left' | 'right' | 'center'
     placeholder?: string
@@ -43,9 +44,12 @@ export enum SurveyType {
 
 export type SurveyQuestion = BasicSurveyQuestion | LinkSurveyQuestion | RatingSurveyQuestion | MultipleSurveyQuestion
 
+export type SurveyQuestionDescriptionContentType = 'html' | 'text'
+
 interface SurveyQuestionBase {
     question: string
     description?: string | null
+    descriptionContentType?: SurveyQuestionDescriptionContentType
     optional?: boolean
     buttonText?: string
     originalQuestionIndex: number
@@ -115,4 +119,6 @@ export interface Survey {
     } | null
     start_date: string | null
     end_date: string | null
+    current_iteration: number | null
+    current_iteration_start_date: string | null
 }